Entry type
The Entry type returns! This time it acts as a tagged union for a serialized
objectclass, storing the objectclass's type along with its proto so we can
automatically deserialize it back to the original enum variant.

This should hopefully ensure that adapters stay decoupled from the ObjectClass
API, and also let us avoid any ObjectClass processing overhead in the event we
just want to return the already-serialized proto back to the client.