Remove log::Id
Instead of having a random log ID, we can use the block ID of the initial block
to identify a log.

This block contains a 16-byte random salt used to derive the encryption key for
the "manager" signing key, as well as the complete ciphertext of a random key,
so together this should be more than enough to ensure every log starts off with
a randomized block ID that won't clash with other logs.