Use rust-protobuf for proto serialization/deserialization
This is a total conversion from the buffoon proto library to rust-protobuf

All protos have been updated to proto3 syntax

Code that was intermingled with the various proto types has been factored
elsewhere for the most part. Some hand edits to the generated proto files still
exist and have hopefully been labeled with comments.

FIXME: objecthash for the "Object" oneof proto type is being calculated
improperly.

Much work still to be done, but this should suffice as an initial pass.

JSON still TBD.